+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ### New features
 
 * OpenStreetMap now available as a map layer (!). Layer style toggle switch between the different Google Maps layer styles (Default, Hybrid, Satellite, Terrain), OpenStreetMap and Wikimedia. (#1181)
-* Use new Google Maps renderer on gms flavour: https://developers.google.com/maps/documentation/android-sdk/renderer
 * OSM map now uses a blue dot and arrow for current location (#1078)
+* Use new Google Maps renderer on gms flavour: https://developers.google.com/maps/documentation/android-sdk/renderer
 * Add preference to draw regions on the map (#1068)
 * Add a preference to prevent location updates any faster than the requested interval (#1168)
+* UI for mode changing is now clearer, using a bottom sheet dialog popup with explanation notes on each mode (#1197)
 * Update to Catalan, Japanese and Polish translations (thanks all!)
 
 ### Bug fixes